js: Rename "compartments" to "realms"
In SpiderMonkey 68, a "realm" is the term for the environment belonging
to a particular global object, so when you enter the "realm" of a global
object you are executing code with that object as the global object.

"Compartments" are still a thing but they are now more focused on
security, and are used for isolating code in different browser tabs in
Firefox, for example. All code in erlang-mozjs is run within the same
compartment.

Some APIs that previously operated on a global object that was passed
in, now operate on the current realm.

Signed-off-by: Peter Lemenkov <lemenkov@gmail.com>